% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignManyDataYears.R
\name{AlignManyDataYears}
\alias{AlignManyDataYears}
\title{AlignManyDataYears}
\usage{
AlignManyDataYears(layer_names)
}
\arguments{
\item{layer_names}{name of the data layer, 
e.g. c('hab_mangrove_extent', 'hab_seagrass_extent', 'hab_saltmarsh_extent')}
}
\value{
Returns an internally used dataframe with the layer data and the data year with the
corresponding scenario year.
#'
}
\description{
A wrapper for the AlignDataYears function when compiling multiple data layers
with the same variable names (e.g., coral, seagrass, mangroves).
}
\examples{
\dontrun{
} 
}
\keyword{ohicore}
