% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateStatusComponent.R
\name{CalculateStatusComponent}
\alias{CalculateStatusComponent}
\title{Compute a single subgoal.}
\usage{
CalculateStatusComponent(DATA, fun, trend.Years = 5, c.name = "year",
  s.name = "region")
}
\arguments{
\item{DATA}{data.frame containing columns 'region', 'value', and (optionally) 'w'}

\item{fun}{(optional) function for calculating the subgoal value, if not specified it will default to a weighted average}

\item{w}{(optional) numeric vector describing the}
}
\value{
stuff
}
\description{
Compute a single subgoal.
}

