% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_plot.R
\name{change_plot}
\alias{change_plot}
\title{change_plot.}
\usage{
change_plot(repo = "ohi-global", scenario = "eez2014",
  commit = "previous", fileSave, save_csv = FALSE, save_png = FALSE)
}
\arguments{
\item{repo}{The repository name, e.g., 'ohi-global'.}

\item{scenario}{The scenario folder name that contains the 'scores.csv' file, e.g., 'eez2014'.}

\item{commit}{The 7 digit sha number identifying the commit, e.g., '4da6b4a'.  Otherwise, it is compared to the previous
commit.}

\item{fileSave}{Name for the figure.  This file will be saved in a folder called 'changePlot_figures'.}

\item{save_csv}{If TRUE, the difference csv file will be saved.}

\item{save_png}{If TRUE, a static png of the image will be saved.}
}
\description{
This function compares OHI scores from the current analysis and a previous commit.  The output is an interactive html plot that is saved
in the working directory in a file called 'changePlot_figures'.
}

