% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_2_rgn.R
\name{collapse_2_rgn}
\alias{collapse_2_rgn}
\title{Collapse region}
\usage{
collapse_2_rgn(df_in, fld_value, fld_id = "rgn_id", collapse_fxn = c("sum",
  "mean", "weighted_mean")[1], collapse_wts = NULL)
}
\arguments{
\item{df_in}{dataset}

\item{fld_value}{field with value, defaults to 'value'}

\item{fld_id}{}

\item{collapse_fxn}{function to collapse duplicate regions into one (example: China, Macau, Hong Kong)}

\item{collapse_wts}{table with weights if weighted mean function is chosen}
}
\description{
Collapse data for duplicate regions.
}
\details{
This function collapses duplicate regions (example: China, Macau, Hong Kong)
}
\keyword{ohi}
