% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_scores_df.R
\name{compare_scores_df}
\alias{compare_scores_df}
\title{Compares scores}
\usage{
compare_scores_df(a_csv, b_csv, r_csv, g_csv)
}
\arguments{
\item{a_csv}{scores.csv for A}

\item{b_csv}{scores.csv for B}

\item{r_csv}{region labels, ie layers/rgn_labels.csv}

\item{g_csv}{goals, ie conf/goals.csv}
}
\description{
Combine two scores.csv files and calculate difference.
}
\details{
Returns a data frame with calculated differences sorted by global (region_id=0), Index, score, goal, dimension, absolute score, is.na(a), is.na(b).
}

