% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\docType{class}
\name{Conf-class}
\alias{Conf-class}
\alias{Conf}
\title{Conf reference class.}
\arguments{
\item{dir}{path to directory containing necessary files, e.g., eez2015/conf}
}
\value{
Object reference class of Config containing:
\itemize{
  \item{\strong{config} - concise description}
  \item{\strong{functions} - concise description}
  \item{\strong{goals} - concise description}
  \item{\strong{pressures_matrix} - concise description}
  \item{\strong{resilience_matrix} - concise description}
  \item{\strong{resilience_categories} - concise description}
  \item{\strong{pressure_categories} - concise description}
}
}
\description{
\code{Conf} object which contains all the configuration information for calculating scores. This creates an R object 
that combines into a single object all the information from the following files: \emph{config.R}, \emph{functions.R}, \emph{goals.csv}, 
\emph{pressures_matrix.csv}, \emph{resilience_matrix.csv}, \emph{resilience_weights.csv}.
}
\details{
To create this object: \code{Conf(dir)}. The \code{dir} is expected to have the following files:
\itemize{
  \item{config.R}
  \item{functions.R}
  \item{goals.csv}
  \item{pressures_matrix.csv}
  \item{resilience_matrix.csv}
  \item{resilience_weights.csv}
  \item{pressure_categories}
}
See also \code{\link{Conf_write}()} to write the configuration back to disk.
}

