% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{georegion_labels}
\alias{georegion_labels}
\title{United Nations Regional Groups}
\format{A data frame with 220 rows and 5 variables:
\itemize{
  \item rgn_id: numeric region id used for Ocean Health Index (1-250)
  \item r0_label: general regional groups (1 group: World)
  \item r1_label: more specific regional groups (7 groups: Africa, Americas, etc.)
  \item r2_label: most specific regional groups (22 groups: Caribbean, Central America, etc.)
  \item rgn_label: region name used for Ocean Health Index (Albania, Angola, etc.)
}}
\usage{
georegion_labels
}
\description{
United Nations geopolitical regions.
In some cases, when data were not available for a country,
we estimated the values using georegional averages at the smallest
possible spatial scale. There are three spatial scales in these data
(general to specific regions: r0_label, r1_label, r2_label).
}
\keyword{datasets}

