% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_scenarios}
\alias{get_scenarios}
\title{Get scenarios}
\usage{
get_scenarios(github_repo, destination_dir, scenarios)
}
\arguments{
\item{destination_dir}{destination directory}

\item{github_url_suffix}{suffix of Github URL in the form 'user/repo'}
}
\description{
Get scenarios from Github.
}
\details{
The scenario files from the containing folder are downloaded and shortcuts specific to R path and OS generated.
}
\keyword{ohi}

