\name{Halpern2012.AO}
\alias{Halpern2012.AO}
\title{Calculate Artisanal Fishing Opportunities.}
\usage{
  Halpern2012.AO(Sao, Oao, PPPpcGDP, ...)
}
\arguments{
  \item{placeholder}{placeholder Sao}

  \item{placeholder}{placeholder Oao}

  \item{placeholder}{placeholder PPPpcGDP}
}
\value{
  1
}
\description{
  Calculate Artisanal Fishing Opportunities.
}

