\name{Halpern2012.CP}
\alias{Halpern2012.CP}
\title{Calculate Coastal Protection}
\usage{
Halpern2012.CP(Cc, Cr, w, A, ...)
}
\arguments{
  \item{placeholder}{placeholder Cc current 'condition' of
  habitat k}

  \item{placeholder}{placeholder Cr reference 'condition'
  of habitat k}

  \item{placeholder}{placeholder A amount of area covered
  by habitat k}

  \item{placeholder}{placeholder w rank weight of habitat
  protective ability}
}
\value{
1
}
\description{
Calculate Coastal Protection
}

