\name{Halpern2012.CS}
\alias{Halpern2012.CS}
\title{Calculate Carbon Storage}
\usage{
Halpern2012.CS(Cc, Cr, A, ...)
}
\arguments{
  \item{placeholder}{placeholder Cc current 'condition' of
  habitat k}

  \item{placeholder}{placeholder Cr reference 'condition'
  of habitat k}

  \item{placeholder}{placeholder A amount of area covered
  by habitat k}
}
\value{
1
}
\description{
Calculate Carbon Storage
}

