\name{Halpern2012.CW}
\alias{Halpern2012.CW}
\title{Calculate Clean Waters.}
\usage{
Halpern2012.CW(a, u, l, d, ...)
}
\arguments{
  \item{placeholder}{placeholder a number of coastal people
  without access to sanitation rescaled to global maximum}

  \item{placeholder}{placeholder u 1 - (nutrient input)}

  \item{placeholder}{placeholder l 1 - (chemical input)}

  \item{placeholder}{placeholder d 1 - (marine debris
  input)}
}
\value{
1
}
\description{
Calculate Clean Waters.
}

