\name{Halpern2012.FP.FIS}
\alias{Halpern2012.FP.FIS}
\title{Calculate Fisheries subgoal of Food Provision.}
\usage{
  Halpern2012.FP.FIS(mMSY, Bt, Tc, ...)
}
\arguments{
  \item{placeholder}{placeholder dBt absolute difference
  between landed biomass and mMSY}

  \item{placeholder}{placeholder mMSY multi-species maximum
  sustainable yield}

  \item{placeholder}{placeholder Tc taxonomic report
  quiality correction factor}

  \item{placeholder}{placeholder Bt wild-caught fishing
  yield}
}
\value{
  1
}
\description{
  Calculate Fisheries subgoal of Food Provision.
}

