\name{Halpern2012.FP.MAR}
\alias{Halpern2012.FP.MAR}
\title{Calculate Mariculture subgoal of Food Provision.}
\usage{
  Halpern2012.FP.MAR(k, Smk, Ac, Yk, ...)
}
\arguments{
  \item{placeholder}{placeholder k each mariculture
  species}

  \item{placeholder}{placeholder Smk sustainability score
  for each species k}

  \item{placeholder}{placeholder Ac area of coastal waters
  (3nm strip)}

  \item{placeholder}{placeholder Yl yield of each species
  k}
}
\value{
  1
}
\description{
  Calculate Mariculture subgoal of Food Provision.
}

