\name{Halpern2012.FP}
\alias{Halpern2012.FP}
\title{Calculate Food Provision.}
\usage{
  Halpern2012.FP(w, dBt, mMSY, Bt, Tc, k, Smk, Ac, Yk, ...)
}
\arguments{
  \item{placeholder}{placeholder k each mariculture
  species}

  \item{placeholder}{placeholder Smk sustainability score
  for each species k}

  \item{placeholder}{placeholder Ac area of coastal waters
  (3nm strip)}

  \item{placeholder}{placeholder Yl yield of each species
  k}
}
\value{
  1
}
\description{
  Calculate Food Provision.
}

