\name{Halpern2012.ICO}
\alias{Halpern2012.ICO}
\title{Calculate Iconic Species subgoal of Sense of Place.}
\usage{
Halpern2012.ICO(S, w, ...)
}
\arguments{
  \item{placeholder}{placeholder S number of assessed
  species in each category}

  \item{placeholder}{placeholder w status weight assigned
  per threat category}
}
\value{
1
}
\description{
Calculate Iconic Species subgoal of Sense of Place.
}

