\name{Halpern2012.LE.ECO}
\alias{Halpern2012.LE.ECO}
\title{Calculate Economies subgoal of Coastal Livelihoods and Economies.}
\usage{
  Halpern2012.LE.ECO(ec, er, ...)
}
\arguments{
  \item{placeholder}{placeholder ec total adjusted revenue
  generated per sector at current time}

  \item{placeholder}{placeholder er total adjusted revenue
  generated per sector at reference time}
}
\value{
  1
}
\description{
  Calculate Economies subgoal of Coastal Livelihoods and
  Economies.
}

