\name{Halpern2012.LE.LIV}
\alias{Halpern2012.LE.LIV}
\title{Calculate Livelihoods subgoal of Coastal Livelihoods and Economies.}
\usage{
Halpern2012.LE.LIV(jc, jr, gc, gr, ...)
}
\arguments{
  \item{placeholder}{placeholder jc total adjusted jobs per
  sector at current time}

  \item{placeholder}{placeholder jr total adjusted jobs per
  sector at reference time}

  \item{placeholder}{placeholder gc average PPP-adjusted
  per-capita annual wages per sector in current region}

  \item{placeholder}{placeholder gr average PPP-adjusted
  per-capita annual wages per sector in reference region}
}
\value{
1
}
\description{
Calculate Livelihoods subgoal of Coastal Livelihoods and
Economies.
}

