\name{Halpern2012.LE}
\alias{Halpern2012.LE}
\title{Calculate Coastal Livelihoods and Economies.}
\usage{
  Halpern2012.LE(jc, jr, gc, gr, ec, er, ...)
}
\arguments{
  \item{placeholder}{placeholder jc total adjusted jobs per
  sector at current time}

  \item{placeholder}{placeholder jr total adjusted jobs per
  sector at reference time}

  \item{placeholder}{placeholder gc average PPP-adjusted
  per-capita annual wages per sector in current region}

  \item{placeholder}{placeholder gr average PPP-adjusted
  per-capita annual wages per sector in reference region}

  \item{placeholder}{placeholder ec total adjusted revenue
  generated per sector at current time}

  \item{placeholder}{placeholder er total adjusted revenue
  generated per sector at reference time}
}
\value{
  1
}
\description{
  Calculate Coastal Livelihoods and Economies.
}

