\name{Halpern2012.LSP}
\alias{Halpern2012.LSP}
\title{Calculate Lasting Special Places subgoal of Sense of Place.}
\usage{
Halpern2012.LSP(CMPA, tCMPA, CP, tCP, ...)
}
\arguments{
  \item{placeholder}{placeholder CMPA coastal marine
  protected area}

  \item{placeholder}{placeholder tCMPA total coastal marine
  area}

  \item{placeholder}{placeholder CP coastline protected}

  \item{placeholder}{placeholder tCP total coastline}
}
\value{
1
}
\description{
Calculate Lasting Special Places subgoal of Sense of Place.
}

