\name{Halpern2012.NP}
\alias{Halpern2012.NP}
\title{Calculate Natural Products. (Needs work)}
\usage{
  Halpern2012.NP(N, wp, Hp, E, R, Nv, Nk, w, ...)
}
\arguments{
  \item{placeholder}{placeholder N number of products that
  have ever been harvested}

  \item{placeholder}{placeholder wp proportional peak
  dollar value of each product relative to the total peak
  dollar value of all products}

  \item{placeholder}{placeholder Hp harvest of a product
  relative to its buffered peak reference point}

  \item{placeholder}{placeholder E exposure term}

  \item{placeholder}{placeholder R risk term}

  \item{placeholder}{placeholder Nv 1 or 2, depending on
  whether or not a viability term is used}

  \item{placeholder}{placeholder Nk number of species in
  each k category of exploitation}

  \item{placeholder}{placeholder w weight assigned to each
  k category of exploitation status}
}
\value{
  1
}
\description{
  Calculate Natural Products. (Needs work)
}

