\name{Halpern2012.SP}
\alias{Halpern2012.SP}
\title{Calculate Sense of Place.}
\usage{
  Halpern2012.SP(S, w, CMPA, tCMPA, CP, tCP, ...)
}
\arguments{
  \item{placeholder}{placeholder S number of assessed
  species in each category}

  \item{placeholder}{placeholder w status weight assigned
  per threat category}

  \item{placeholder}{placeholder CMPA coastal marine
  protected area}

  \item{placeholder}{placeholder tCMPA total coastal marine
  area}

  \item{placeholder}{placeholder CP coastline protected}

  \item{placeholder}{placeholder tCP total coastline}
}
\value{
  1
}
\description{
  Calculate Sense of Place.
}

