\name{Halpern2012.TR}
\alias{Halpern2012.TR}
\title{Calculate Tourism and Recreation.}
\usage{
  Halpern2012.TR(D, t, V, S, ...)
}
\arguments{
  \item{placeholder}{placeholder D number of tourist-days}

  \item{placeholder}{placeholder t most recent year}

  \item{placeholder}{placeholder V total region population
  size}

  \item{placeholder}{placeholder S sustainability factor}
}
\value{
  1
}
\description{
  Calculate Tourism and Recreation.
}

