% Generated by roxygen2 (4.0.1): do not edit by hand
\name{launch_app}
\alias{launch_app}
\title{Launch the browser application}
\usage{
launch_app(dir_scenario = getwd(), debug = F, launch.browser = T, ...)
}
\arguments{
\item{dir_scenario}{path to the scenario directory. defaults to working directory (\code{\link{getwd}}).}

\item{...}{arguments passed to \code{shiny::runApp}}
}
\description{
This function launches the OHI application into a web browser with the scenario data laoded.
}
\details{
The dir_scenario is expected to have the following files and directories (*/), some of which are optional:
\itemize{
  \item{\emph{conf} - configuration directory. see \code{\link{Conf}} for details.}
  \item{\emph{layers}, \emph{layers.csv} - layers.csv registry and layers directory. see \code{\link{Layers}} for details.}
  \item{\emph{spatial} - spatial directory, containing the regions_gcs.js. TODO: documentation on this.}
  \item{\emph{scores}(optional) - scores output from  \code{\link{CalculateAll}}.}
}
}
\examples{
\dontrun{
launchApp('~/ohi-global/eez2013')
}
}
\keyword{app}

