% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_cmp.R
\name{launch_cmp}
\alias{launch_cmp}
\title{Launch the scneario comparison browser application}
\usage{
launch_cmp(launch.browser = T, port = NULL, display.mode = "normal", ...)
}
\description{
This function launches the OHI Scenario Comparison application into a web browser.
You'll need to have the https://github.com/ropensci/git2r library installed.
}

