% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\docType{class}
\name{Layers-class}
\alias{Layers-class}
\alias{Layers}
\title{Layers reference class.}
\arguments{
\item{layers.csv}{path to comma-seperated value file with a row of metadata for each dataset 
(or layer) used in OHI analysis.}

\item{layers.dir}{full path to the directory containing the layers files 
(the csv files that correspond to each entry in layers.csv).}
}
\value{
Object (non-instantiated) reference class, containing
\itemize{
 \item{\strong{meta} - metadata data frame of original layers.csv}
 \item{\strong{data} - named list of data frames, one per layer}
 \item{\strong{targets} - a named list of character vectors indicating a layer's targets, 
 goal (status, trend) or dimension (pressures, resilience)}
}
}
\description{
\code{Layers} object which pulls together all the data layers for calculating scores, 
along with their metadata.
}
\details{
This function creates an R object that combines into a single object all the information 
from the layers files in layers.dir and the layers.csv metadata. Individual layers can be 
accessed as: layer_object_name$data$layer_name. To create this object: \code{Layers(dir)}. 
The \code{layers.dir} should contain all the csv filenames listed in the \code{layers.csv} 
file and is expected to have the following columns:
\itemize{
  \item{\strong{layer} - unique layer identifier (no spaces or special characters)}
  \item{\strong{targets} - a space delimited list of targets (goal code, 'Pressures', 
  'Resilience' or 'Regions') for which this layer is applied}
  \item{\strong{name} - name of the variable}
  \item{\strong{description} - detailed description}
  \item{\strong{units} - units of the value}
  \item{\strong{citation} - reference for documentation, typically a heading code for 
  a supplemental document}
  \item{\strong{filename} - the csv data file for the layer}   
  \item{\strong{fld_value} - required field in the layer csv file containing the value, 
  which is often best named as a shorthand for the units without spaces or special characters}
}
The layers.dir directory should contain all csv filenames listed in the layers.csv file.
}

