% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{layers.Global2012.Nature2012ftp}
\alias{layers.Global2012.Nature2012ftp}
\title{Layers accompanying Nature 2012 publication on the FTP site for Global 2012 analysis.}
\format{a \code{\link{Layers}} object}
\usage{
layers.Global2012.Nature2012ftp
}
\description{
These layers get used to calculate the Ocean Health Index.
}
\references{
\url{http://ohi-science.org}
}
\keyword{datasets}

