% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_eez_base_updater.R
\name{layers_eez_base_updater}
\alias{layers_eez_base_updater}
\title{Update layers_eez_base.csv
By Peter Menzies}
\usage{
layers_eez_base_updater()
}
\description{
Convenient interface for changing directory paths of recently updated layers in layers_eez_base.csv
No arguments required - user will be prompted for needed information
}
\keyword{ohi}
