% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvalues.R
\name{mapvalues}
\alias{mapvalues}
\title{mapvalues}
\usage{
mapvalues(x, from, to, warn_missing = TRUE)
}
\arguments{
\item{x}{the factor or vector to modify}

\item{from}{a vector of the items to replace}

\item{to}{a vector of replacement values}

\item{warn_missing}{print a message if any of the old values are not actually present in x}
}
\value{
Returns a vector with new values.
}
\description{
Replace specified values with new values, in a vector or factor. This is copied from plyr.
}
\details{
#' If x is a factor, the matching levels of the factor will be replaced with the new values.

The related revalue function works only on character vectors and factors, but this function works 
on vectors of any type and factors.
}
\examples{
\dontrun{
} 
}
\keyword{ohicore}
