% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_2_rgn.R
\name{name_2_rgn}
\alias{name_2_rgn}
\title{Name to region (New version: doesn't collapse data)}
\usage{
name_2_rgn(df_in, fld_name = "country", flds_unique = NULL,
  keep_fld_name = TRUE)
}
\arguments{
\item{df_in}{dataset}

\item{fld_name}{field name of the region from the dataset}

\item{flds_unique}{unique identifying fields for the dataset, e.g., c("rgn_name", "year")}

\item{keep_fld_name}{keep original name}
}
\description{
Convert country names to ohi region ids.
}
\details{
This function translates name to region id with a lookup.
}
\keyword{ohi}

