% Generated by roxygen2 (4.0.2): do not edit by hand
\name{name_to_rgn}
\alias{name_to_rgn}
\title{Get scenarios}
\usage{
name_to_rgn(d, fld_name = "country", flds_unique = fld_name,
  fld_value = "value", collapse_fxn = c("sum_na", "mean",
  "weighted.mean")[1], collapse_csv = NULL, collapse_flds_join = NULL,
  dir_lookup = "../ohiprep/src/LookupTables",
  rgn_master.csv = file.path(dir_lookup, "eez_rgn_2013master.csv"),
  rgn_synonyms.csv = file.path(dir_lookup, "rgn_eez_v2013a_synonyms.csv"),
  add_rgn_name = F, add_rgn_type = F)
}
\arguments{
\item{d}{dataset}

\item{fld_name}{field name of the region from the dataset}

\item{flds_unique}{field name for the dataset}

\item{fld_value}{field with value, defaults to 'value'}
}
\description{
Get scenarios from Github.
}
\details{
This function translates name to region id with a lookup.

TEMPORARY. More notes here on documenting and including new functions in ohicore.
See http://r-pkgs.had.co.nz/man.html for function documentation details.
TODO: finish out parameters above.
FINISH: After any documentation edits or a new function file added, you'll need to do the following:
library(devtools)
document() # updates DESCRIPTION and NAMESPACE files of ohicore package
# test() # might bonk, but can still build usually
build()

commit, push, done.
}
\keyword{ohi}

