% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_git_csv}
\alias{read_git_csv}
\title{Read CSV from local Git repository}
\usage{
read_git_csv(repo, hex, path, ...)
}
\arguments{
\item{repo}{path to local repository}

\item{hex}{SHA hex of commit}

\item{path}{to csv file with the repository as root}
}
\description{
Read CSV from local Git repository.
}
\examples{
\dontrun{
# get csv from github repository by SHA hex of commit
d = read_git_csv('~/github/ohi-global', 'a81a8213', 'scores.csv')
head(d)
}
}
\keyword{git}

