% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgn_eez_area.R
\docType{data}
\name{rgn_eez_area}
\alias{rgn_eez_area}
\title{Region eez areas}
\format{A data frame with 236 rows and 2 variables:
\describe{
  \item{rgn_id}{region id used for Ocean Health Index (starting with 2013 assessment)}
  \item{area_km2}{area of region's eez}
}}
\usage{
rgn_eez_area
}
\description{
Area (km2) of each OHI region's eez.
}
\keyword{datasets}
