% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_check.R
\name{score_check}
\alias{score_check}
\title{score_check}
\usage{
score_check(scenario_year, commit = "previous", file_name,
  save_csv = FALSE, save_png = FALSE, NA_compare = TRUE)
}
\arguments{
\item{scenario_year}{if there are multiple scenario years with an assessment, choose the year of interest}

\item{commit}{commit that the current scores are compared to ('previous' for last commit or 7 digit SHA, e.g., '4da6b4a')}

\item{file_name}{descriptive name used to save file outputs}

\item{save_csv}{TRUE/FALSE, to save a csv file of the data}

\item{save_png}{TRUE/FALSE, save static png file of interactive plot}

\item{NA_compare}{TRUE/FALSE, compares the NA values between the datasets}
}
\value{
Returns an interactive html plot of the change in scores (current score - previous score) for all score 
        dimensions and goals/subgoals.  Other outputs can also be saved.
#'
}
\description{
Used to error check scores after addition of new data by comparing current scores to a previous commit. 
Outputs are saved in a file within the github repository called "check scores". 
 Outputs include 1) an interactive html plot to examine the change 
(current score minus previous score) in all score dimensions (status, trend, resilience, pressure, future) 
across all goals/subgoals; 2) a static png plot of the interactive html plot; 3) a csv file with the 
current and previous scores; 4) a comparison of the NA scores
}
\examples{
\dontrun{
} 
}
\keyword{ohicore}
