\docType{data}
\name{scores.Global2012.www2013}
\alias{scores.Global2012.www2013}
\title{Scores resulting from the 2013 web launch applied to Global 2012 analysis.}
\format{a \code{\link{Scores}} object}
\usage{
scores.Global2012.www2013
}
\description{
These scores are the results of the Ocean Health Index.
}
\references{
\url{http://ohi-science.org}
}
\keyword{datasets}

