\docType{data}
\name{scores.Global2013.www2013}
\alias{scores.Global2013.www2013}
\title{Scores resulting from the 2013 web launch applied to Global 2013 analysis.}
\format{a \code{\link{Scores}} object}
\description{
  These scores are the results of the Ocean Health Index.
}
\references{
  \url{http://ohi-science.org}
}
\keyword{datasets}

