% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sovregion_labels.R
\docType{data}
\name{sovregion_labels}
\alias{sovregion_labels}
\title{Territories}
\format{
A data frame with 220 rows and 5 variables:
\describe{
  \item{rgn_id}{numeric region id used for Ocean Health Index (1-250)}
  \item{r0_label}{sovreign country name}
  \item{r1_label}{sovreign country name}
  \item{r2_label}{sovreign country name}
  \item{rgn_label}{region name used for Ocean Health Index (Albania, Angola, etc.)}
}
}
\usage{
sovregion_labels
}
\description{
Identifies a country's territories.
}
\details{
This information is used to assign data to a country's territories.
}
\keyword{datasets}
