% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/sovregions.R
\docType{data}
\name{sovregions}
\alias{sovregions}
\alias{sovregions}
\title{Territories (streamlined)}
\format{A data frame with 220 rows and 5 variables:
\describe{
  \item{rgn_id}{region id used for Ocean Health Index (1-250)}
  \item{r0}{sovreign country region id}
  \item{r1}{sovreign country region id}
  \item{r2}{sovreign country region id}
  \item{fld_wt}{identifies territory countries (0) and sovreign regions (1)} 
}}
\usage{
sovregions

sovregions
}
\description{
This information is used to assign data to a country's territories.
(This is a version of the sovregion_labels data)

This information is used to assign data to a country's territories.
(This is a version of the sovregion_labels data)
}
\keyword{datasets}
