% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_git_csv_value.R
\name{trace_git_csv_value}
\alias{trace_git_csv_value}
\title{Trace Value from CSV through history of local Git repository}
\usage{
trace_git_csv_value(repo, csv, subset_str, select, verbose = T)
}
\arguments{
\item{repo}{path to repository on local filesystem}

\item{csv}{path to csv file with the repository as root}

\item{subset_str}{subset argument to the function \code{\link{subset}} quoted as string to extract row of data from csv}

\item{select}{field to select from subsetted row}
}
\value{
data.frame having columns: hex, when, message, v.
}
\description{
Trace Value from CSV through history of local Git repository
}
\details{
If you have trouble running this function, please make sure:
1) your path resolves to a local Git repository, 
2) you have the latest git2r (try \code{devtools::install_github('ropensci/git2r')}).
}
\examples{

\dontrun{
# trace the value for a csv from github repository
d = trace_git_csv_value('~/github/ohicore', 'inst/extdata/scores.Global2013.www2013.csv', "goal=='ECO' & dimension=='status' & region_id==237", 'score')
head(d)
}

}
\keyword{git}
