% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_shortcuts}
\alias{ReadScenario}
\alias{WriteScenario}
\alias{write_shortcuts}
\title{Write shortcuts}
\usage{
write_shortcuts(dir_scenario = getwd(), all_os = F)
}
\arguments{
\item{dir_scenario}{per \code{\link{launch_app}}}

\item{all_os}{for all operating systems, ie Mac if on Windows, or Windows if on Mac}
}
\description{
Write shortcuts to launch the application.
}
\keyword{ohi}

