% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteRefPoint.R
\name{WriteRefPoint}
\alias{WriteRefPoint}
\title{WriteRefPoint}
\usage{
WriteRefPoint(goal, method, ref_pt)
}
\arguments{
\item{goal}{goal name (e.g., "FIS")}

\item{method}{method of identifying reference point (e.g., "95th quantile among regions")}

\item{year}{default is the scenario year}

\item{reference_point}{reference point value}
}
\value{
rbinds provided information to the temp/reference_pts.csv so this can be 
referenced later on
#'
}
\description{
Writes information about reference points to a csv file located in temp/reference_pts.csv
}
\examples{
\dontrun{
} 
}
\keyword{ohicore}
