% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CalculateAll}
\alias{CalculateAll}
\title{Calculate All}
\usage{
CalculateAll(conf, layers, debug = F)
}
\arguments{
\item{conf}{of class \code{\link{Conf}}}

\item{layers}{of class \code{\link{Layers}}}

\item{debug}{print debug messages (default=FALSE)}
}
\value{
Returns a data.frame of scores having the following columns:
\itemize{
  \item{\emph{region_id} - unique numeric region identifier, reserving 0 as the region_id for the area-weighted average of the entire study area}
  \item{\emph{goal} - the goal code or Index}
  \item{\emph{dimension} - the dimension code, one of: status, trend, pressures, resilience, future, score}
  \item{\emph{score} - the numeric score: 0-100 for all dimensions, except trend (-1 to 1)}
}
}
\description{
Calculate all scores, given layers and configuration.
}
\details{
Performs the following sequence of functions, some of which are [optional]:
\enumerate{
  \item [functions.R:\code{Setup}()] - execute function \code{Setup}() if defined in file functions.R.
  This function typically installs extra packages upon which the other functions in functions.R depend.
  \item \code{\link{CalculatePressuresAll}()} - calculate pressures across all goals using pressures_matrix.csv.
  \item \code{\link{CalculateResilienceAll}()} - calculate resilience across all goals using resilience_matrix.csv and resilience_weights.csv.
  \item goals.csv:\code{preindex_functions} - execute code in the \code{preindex_function} column of the goals.csv file
  based on \code{order_calculate} using functions defined in functions.R.
  These funcutions are usually for calculating the goal's status and trend dimensions, ie
  the additional dimensions beyond pressures and resilience needed to calculate a goal index score.
  \item \code{\link{CalculateGoalIndex}()} - run function for every goal having a status dimension assigned from the \code{preindex_functions}.
  \item goals.csv:\code{postindex_functions} - execute code in the \code{postindex_function} column of the goals.csv file based on \code{order_calculate}
  using functions defined in functions.R. These functions are usually for goals containing subgoals, ie those without their own directly calculated
  index scores, but rather scores representing averages of subgoals.
  \item regional index - calculate regional index score as weighted mean using goals.csv:\code{weight}.
  \item regional likely future - calculate regional likely future score (ie goal='Index' and dimension='future') across supragoals (ie goals without a \code{parent} in goals.csv).
  \item [functions.R:\code{PreGlobalScores}()] - execute function \code{PreGlobalScores}() if defined in file functions.R.
  This function could perform a variety of operations on the regional scores, strategically before calculating the global scores.
  \item global (region_id=0) scores - calculate scores for global (region_id=0) with regional values weighted by config.R:\code{layer_region_areas}.
  \item [functions.R:\code{FinalizeScores}()] - execute function \code{FinalizeScores}() if defined in file functions.R.
  This function could perform a variety of operations on the regional and global scores.
}
}
\examples{
\dontrun{
## run a scenario assuming setwd() to directory containing default names for directories and files
## setup
require(ohi)
conf       = Conf('conf')
layers     = Layers(layers.csv = 'layers.csv',
                    layers.dir = 'layers')
## calculate
scores = CalculateAll(conf, layers, debug=T)

## write
write.csv(scores, 'scores.csv', na='', row.names=F)
}
}
\keyword{ohi}

