% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePressuresAll.R
\name{CalculatePressuresAll}
\alias{CalculatePressuresAll}
\title{Calculate all the pressures score for each (sub)goal.}
\usage{
CalculatePressuresAll(layers, conf)
}
\arguments{
\item{layers}{object \code{\link{Layers}}}

\item{conf}{object \code{\link{Conf}}}
}
\value{
data.frame containing columns 'region_id' and per subgoal pressures score
}
\description{
Calculate all the pressures score for each (sub)goal.
}
