% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CalculatePressuresAll}
\alias{CalculatePressuresAll}
\title{Calculate all the pressures score for each (sub)goal.}
\usage{
CalculatePressuresAll(layers, conf, gamma = 0.5, debug = F)
}
\arguments{
\item{layers}{object \code{\link{Layers}}}

\item{conf}{object \code{\link{Conf}}}

\item{gamma}{(optional) if not specified defaults to 0.5}
}
\value{
data.frame containing columns 'region_id' and per subgoal pressures score
}
\description{
Calculate all the pressures score for each (sub)goal.
}

