% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CalculateResilienceAll}
\alias{CalculateResilienceAll}
\title{Calculate all the resilience score for each (sub)goal.}
\usage{
CalculateResilienceAll(layers, conf, debug = FALSE)
}
\arguments{
\item{layers}{object \code{\link{Layers}}}

\item{conf}{object \code{\link{Conf}}}
}
\value{
data.frame containing columns 'region_id' and per subgoal resilience score
}
\description{
Calculate all the resilience score for each (sub)goal.
}

