% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Conf.R
\docType{class}
\name{Conf-class}
\alias{Conf}
\alias{Conf-class}
\title{Conf reference class.}
\arguments{
\item{dir}{path to directory containing necessary files}
}
\value{
object reference class of Config containing:
\itemize{
  \item{\emph{config}}
  \item{\emph{functions}}
  \item{\emph{goals}}
  \item{\emph{pressures_matrix}}
  \item{\emph{resilience_matrix}}
  \item{\emph{resilienceweights}}
}
}
\description{
Conf reference class.
}
\details{
To create this object, \code{Conf(dir)}. The \code{dir} is expected to have the following files:
\itemize{
  \item{\emph{config.R}}
  \item{\emph{functions.R}}
  \item{\emph{goals.csv}}
  \item{\emph{pressures_matrix.csv}}
  \item{\emph{resilience_matrix.csv}}
  \item{\emph{resilienceweights.csv}}
}
See also \code{\link{Conf_write}()} to write the configuration back to disk.
}

