% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_scenarios.R
\name{get_scenarios}
\alias{get_scenarios}
\title{Get scenarios}
\usage{
get_scenarios(github_repo, destination_dir)
}
\arguments{
\item{destination_dir}{destination directory}

\item{github_url_suffix}{suffix of Github URL in the form 'user/repo'}
}
\description{
Get scenarios from Github.
}
\details{
The scenario files from the containing folder are downloaded and shortcuts specific to R path and OS generated.
}
\keyword{ohi}

