% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_scenarios}
\alias{get_scenarios}
\title{Get scenarios}
\usage{
get_scenarios(github_url_suffix, destination_dir)
}
\arguments{
\item{github_url_suffix}{suffix of Github URL in the form 'user/repo'}

\item{destination_dir}{destination directory}
}
\description{
Get scenarios from Github.
}
\details{
The scenario files from the containing folder are downloaded and shortcuts specific to R path and OS generated.
}
\keyword{ohi}

