\docType{data}
\name{layers.Global2012.www2013}
\alias{layers.Global2012.www2013}
\title{Layers used for the 2013 web launch applied to Global 2012 analysis.}
\format{a \code{\link{Layers}} object}
\description{
  These layers get used to calculate the Ocean Health
  Index.
}
\references{
  \url{http://ohi-science.org}
}
\keyword{datasets}

