% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_git_csv.R
\name{read_git_csv}
\alias{read_git_csv}
\title{Read CSV from local Git repository}
\usage{
read_git_csv(repo, hex, path, ...)
}
\arguments{
\item{repo}{path to local repository}

\item{hex}{SHA hex of commit}

\item{path}{to csv file with the repository as root}
}
\description{
Read CSV from local Git repository.
}
\details{
If you have trouble running this function, please make sure:
1) your path resolves to a local Git repository,
2) the hex is valid (see history in RStudio Git or on Github) and
3) you have the latest git2r (try \code{devtools::install_github('ropensci/git2r')}).
}
\examples{

\dontrun{
# get csv from github repository by SHA hex of commit
d = read_git_csv('~/github/ohi-global', 'a81a8213', 'scores.csv')
head(d)
}

}
\keyword{git}

