% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores.Global2013.www2013}
\alias{scores.Global2013.www2013}
\title{Scores resulting from the 2013 web launch applied to Global 2013 analysis.}
\format{a \code{\link{Scores}} object}
\usage{
scores.Global2013.www2013
}
\description{
These scores are the results of the Ocean Health Index.
}
\references{
\url{http://ohi-science.org}
}
\keyword{datasets}

