% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_to_geojson.R
\name{shp_to_geojson}
\alias{shp_to_geojson}
\title{Create GeoJSON from Shapefile}
\usage{
shp_to_geojson(shp, js, geojson = sprintf("\%s.geojson",
  tools::file_path_sans_ext(js)))
}
\arguments{
\item{shp}{path to shapefile with .shp extension, needs rgn_id and rgn_name fields}

\item{js}{path to output javascript file with variable 'regions' of geojson content}

\item{geojson}{path to output GeoJSON file. defaults to *.geojson of *.js file.}
}
\description{
Create GeoJSON file needed for interactive map in Shiny app
}
\details{
Uses rgdal to write GeoJSON.
}
\keyword{geojson}
\keyword{shapefile}
