% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialSchemes.R
\docType{class}
\name{SpatialSchemes-class}
\alias{SpatialSchemes}
\alias{SpatialSchemes-class}
\title{SpatialSchemes reference class.}
\value{
object (non-instantiated) reference class of SpatialSchemes
}
\description{
SpatialSchemes reference class.
}

