% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{SpatialSchemes-class}
\alias{SpatialSchemes}
\alias{SpatialSchemes-class}
\title{SpatialSchemes reference class.}
\value{
object (non-instantiated) reference class of SpatialSchemes
}
\description{
SpatialSchemes reference class.
}

