% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_shortcuts.R
\name{write_shortcuts}
\alias{ReadScenario}
\alias{WriteScenario}
\alias{write_shortcuts}
\title{Write shortcuts}
\usage{
write_shortcuts(dir_scenario = getwd(), os_files = 1)
}
\arguments{
\item{dir_scenario}{per \code{\link{launch_app}}}

\item{os_files}{has possible values 0,1 or 2 to determine the files output. See Details.}
}
\description{
Write shortcuts to launch the application.
}
\details{
The following files are generated based on the value of \code{os_files}:
\itemize{
  \item{\emph{0}: the files not dependant on operating system are output: launch_app_code.R, calculate_scores.R.}
  \item{\emph{1}: in addition to 0, the shortcut for only the running operating system is output: launch_app.bat for Windows, launchApp.command for Mac.}
  \item{\emph{2}: in addition to 0, shortcuts for both operating systems are output: launch_app.bat for Windows, launchApp.command for Mac.}
}
If this function with the \code{os=2} argument is run on a Mac, then the launch_app.bat is unlikely to match the R path on a Windows machine.
Whereas, the launchApp.command shortcut on a Mac should still work even if generated on a Windows machine.
}
\keyword{ohi}

